requirements.txt
folium==0.14.0
inflection==0.5.1
pandas==1.5.3
Pillow==9.4.0
plotly==5.14.0
streamlit==1.20.0
streamlit-folium==0.11.1